pyproject.toml
[tool.poetry]
name = "brotein"
version = "0.1.0"
description = ""
authors = ["catvin desktop <b02202040@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
fastapi = {extras = ["all"], version = "^0.114.1"}
pydantic = "^2.9.1"
sqlalchemy = "^2.0.34"
passlib = "^1.7.4"
pyjwt = "^2.9.0"
pymysql = "^1.1.1"
python-dotenv = "^1.0.1"
bcrypt = "4.0.1"
openai = "^1.44.1"
uvicorn = "^0.30.6"
pytz = "^2024.2"
boto3 = "^1.35.22"
pillow = "^10.4.0"
matplotlib = "^3.9.2"
opencv-python = "^4.10.0.84"
ultralytics = "^8.3.15"


[tool.poetry.group.dev.dependencies]
setuptools = "^74.1.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
